refactor: move markdown-table into core/lib/actions and reuse parseRulesOrThrow in report/src
markdownTable is a generic GitHub-flavored markdown table renderer with
no report-specific knowledge (see its own docstring) — move it next to
annotation.ts under core/lib/actions, since both exist to format text
for GitHub Actions' output surfaces (annotations, Job Summary tables).

Also switch report/src/main.ts's known_blocked_rules parsing from the
raw core/shared/lib/rules.js call to the shared parseRulesOrThrow (see
core/lib/acl/rules.ts), matching setup/src/main.ts and run/src/main.ts
instead of duplicating the same validation call independently.
